refactor: restructured project folder layout
